+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+
+### Fixed
+
+- Fixed `make_mapped_structure` to properly handle atom properties (i.e. forces) with implied vacancies.
+
+
 ## [v2.0a5] - 2024-08-18
 
 ### Added